pyproject.toml
[project]
name = "expense-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.6.0",
    "fastapi>=0.115.14",
    "litellm>=1.74.0.post1",
    "openai-agents>=0.1.0",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
]
